Add files via upload
This is the HTML, minimal CSS, and script for a dice roller. It takes the quantity, number of sides on the die, and a modifier to give you the total for the role. In this version, the preventDefault and onclick functions are not working properly and are currently being edited. This is still a work in progress.